v1: don't set GOPROXY=direct
Initially we had used GOPROXY=direct when rewriting the go.mod in order
to ensure that our potentially very-new commits we were targeting with
the replace directives were fetched correctly. However, when this job is
running normally, it won't execute more than once a day and unsetting
the GOPROXY settings is only necessary when trying to resolve commits
that are O(minutes) new.

We found that using GOPROXY=direct exposed us to missing modules that
otherwise would build normally since the proxy servers cache everything
exactly for this purpose.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>