Add support to limit the what gets downstreamed in v0
The merge tool always pulls the master branch from all the staging repos.
This is a problem if one of the dependencies has been updated to a version
of something that is incompatible with the base repository, and the base
repository still references an older dependency.

For instance: operator-lifecycle-manager may reference older verisons
of operator-framework/api and operator-framework/operator-registry.

If one of those dependencies is updated to a more recent version of say,
k8s, but OLM hasn't pulled it in, that will cause merge problems.

What should happen is upstream OLM should integrate with those newer
dependencies before they're pulled downstream. Until then, the dependencies
should remain at the versions listed in OLM's depdencies, as specified in
go.mod.

This is done by looking at the latest OLM and seeing what versions
of `api` and `operator-registry` should be referenced.

In addition to the above:

Remove the `go mod` commands run in staging, those commands are not
necessary as upstream should be running them and keeping their go.*
files maintained.

Add logic to handle commits that have been downstreamed after the
tagged version that OLM expects. This is mostly for the transision
period to limiting downstreaming by tag.

Signed-off-by: Todd Short <todd.short@me.com>