Calculate v0 `centeralRef` up front
When the bumper is run as a job, `-central-ref=FETCH_HEAD` is set.
When the central-ref is used, the value is now different due to the
new `calculateRepoRefs` functionality (which fetches).

The default value is `origin/master`, but the job's git configuration
does not allow for that value; it's set to FETCH_HEAD.

So, calculate the value of central-ref before we do anything.

Signed-off-by: Todd Short <todd.short@me.com>