Handle go.mod conflicts
Restore running `go mod` commands in the staging directory.

And then, due to the `go mod` commands running in the staging directory,
the `go.mod` file may have conflicts. So resolve it as "ours", and then
use the `go mod` commands in the staging directory to update it.

The conflicts can arise due to downstream-only code in a staging
directory affecting the `go mod` command results.

Signed-off-by: Todd Short <todd.short@me.com>