Handle manifest generation better
Manifests are generated by a commit via the bumper. However, new commits
may attempt to modify the manifests. When the bumper comes along to
synchronize downstream, these commits are cherry-picked before the
manifest generation commit. This means that there will be a conflict in
the manifests.

To work around this, we generate the manifests in openshift in every
carry commit. This also requires running the `go mod` commands to
retreive the code necessary, for every carry commit.

Subsequently, the manifest generation and `go mod` commands are no
longer run in openshift after the sequence of cherry picks.

This allows downstreaming to occur with freshly generated manifests,
and allows downstream-only commits to modify the manifests as required.

Signed-off-by: Todd Short <todd.short@me.com>