Add ugly pause option for manual intervention
It seems that some ART commits may be modifying the `go.mod` file,
specifically the go version and toolchain lines. This is updated
based on the go version used by ART, and it may differ from the
upstream version. This can cause subsequent downstream merges to
fail due to cherry-pick conflicts.

Add `--pause-on-cherry-pick-error` option for use when manually
running the bumper. The process of the bumper is complex enough
such that it is easy to make a mistake when attempting to complete
a broken downsteeam. This option will pause the bumper to allow
the user to fix the cherry-pick so that the bumper can continue.

Signed-off-by: Todd Short <todd.short@me.com>