OPRUN-3741: OLMv0: Use main branch when it's available
CRT is renaming downstream branches from master to main.
This takes into account the current downstream branch name to use for the PR.
Assume "main", but change to "master" for OLMv0 if "main" is not found.

It also updates other uses of "master" to be clear as to the use (i.e. upstream)

Signed-off-by: Todd Short <todd.short@me.com>