Add cherry-pick merge strategy
Update the merge strategy to be `ort`, which is the default, but with the
option of `ours`, which will prefer the incoming cherry-pick changes over
the existing code. This is only done during the downstream cherry-picks
which ought to be safe.

The conflict might come from go.mod or go.sum files that were updated
by a prior `go mod tidy` or `go mod vendor` command.

One of the last drop commits is a `go mod tidy` and `go mod vendor`
over _all_ the vendored code (basically wherever a `go.mod` is found).

Signed-off-by: Todd Short <todd.short@me.com>